🤖 Refactor: terraformConvertor 코드 분리
